+++ src/gamspy/_symbols/implicits/implicit_parameter.py
@@ -99,7 +99,7 @@ class ImplicitParameter(ImplicitSymbol, operable.Operable):
         self.container._add_statement(statement)
         self.parent._assignment = statement
 
-        self.container._synch_with_gams()
+        self.container._synch_with_gams(gams_to_gamspy=True)
         self.parent._winner = "gams"
 
     def __eq__(self, other):  # type: ignore