+++ src/gamspy/_symbols/parameter.py
@@ -244,7 +244,9 @@ class Parameter(gt.Parameter, operable.Operable, Symbol):
             if records is not None:
                 self.setRecords(records, uels_on_axes=uels_on_axes)
             else:
-                self.container._synch_with_gams()
+                self.container._synch_with_gams(
+                    gams_to_gamspy=self._is_miro_input
+                )
 
             container._options.miro_protect = previous_state
 
@@ -284,7 +286,7 @@ class Parameter(gt.Parameter, operable.Operable, Symbol):
         self.container._add_statement(statement)
         self._assignment = statement
 
-        self.container._synch_with_gams()
+        self.container._synch_with_gams(gams_to_gamspy=True)
         self._winner = "gams"
 
     def __eq__(self, other):  # type: ignore
@@ -426,7 +428,7 @@ class Parameter(gt.Parameter, operable.Operable, Symbol):
         """
         super().setRecords(records, uels_on_axes)
 
-        self.container._synch_with_gams()
+        self.container._synch_with_gams(gams_to_gamspy=self._is_miro_input)
         self._winner = "python"
 
     def gamsRepr(self) -> str: