+++ tests/unit/test_options.py
@@ -120,7 +120,7 @@ def test_options(data):
 
     assert (
         content
-        == 'limcol = "0"\nlimrow = "0"\nsolprint = "0"\nsolvelink = "2"\npreviouswork = "1"\ntraceopt = "3"\ngdxSymbols = "newOrChanged"\nlogoption = "0"'
+        == 'limcol = "0"\nlimrow = "0"\nsolprint = "0"\nsolvelink = "2"\npreviouswork = "1"\ntraceopt = "3"\nlogoption = "0"'
     )